trouble shooting

+++ fhaa/views/request_views.py
@@ -42,6 +42,8 @@ def req_post() :
     for hospital in hospitals:
         # 환자 <-> 병원 거리
         distance = gps.compare((location_lat, location_lon), (hospital.hos_lat, hospital.hos_lnt))
+        if distance < 1.:
+            distance = 0.            
         r = Request(
                 req_type = f_req_type, 
                 req_loc = addr_now,
@@ -78,12 +80,18 @@ def board():
             f_req_id = request.form.get('req_id')
             request_ = Request.query.filter_by(req_id=f_req_id)
             
+            # request_.first()가 None 경우 == 받은 req_id의 데이터가 이미 삭제되었을 경우
+            if request_.first() == None:
+               error = "이미 만료된 요청입니다."
+               return redirect(url_for('request.board'))
+            
             # 현재 시간과 req_date+req_time을 비교하여 현재 시간이 더 크면 불가해야함     
             # 병원이 새로고침을 하지 않은 상태라면 `진료 요청시간`을 넘겼음에도 수락이 되기 때문에
             # 아래와 같이 현재 시간과 비교하여 처리함
-            req_time = datetime.timedelta(minutes=request_.first().req_time)
+            req_time = datetime.timedelta(minutes=int(request_.first().req_time))
             req_limit_time = request_.first().req_date + req_time   
             if datetime.datetime.now() > req_limit_time:
+                # TODO : 사용자에게 알려야 함
                 error = "이미 진료 요청시간을 초과하였습니다."
             else:
                 request_.update(dict(req_chk=0))
@@ -93,10 +101,12 @@ def board():
         else:
             f_req_id = request.form.get('req_id')
             request_ = Request.query.filter_by(req_id=f_req_id)
+            # request_.first()가 None 경우 == 받은 req_id의 데이터가 이미 삭제되었을 경우
+            if len(request_) == 0:
+                return redirect(url_for('request.board'))
             db.session.delete(request_.first())
             db.session.commit()
             
-        flash(error)
         return redirect(url_for('request.board'))
     
     page = request.args.get('page', type=int, default=1)  # 페이지
